+++ src/types/contextMenu/CommunityContextMenuOption.ts
@@ -1,0 +1,6 @@
+export type CommunityContextMenuOption =
+  | 'block'
+  | 'info'
+  | 'modlog'
+  | 'moderators'
+  | 'bannedUsers';